requirements.txt
opencv_python
setuptools
PySide6>=6.8.0
dayu-path>=0.5.2
msgpack>=1.1.0
requests>=2.31.0
pdfplumber>=0.11.5
torch>=2.6.0
torchvision>=0.21.0
py7zr>=0.20.8
rarfile>=4.1
img2pdf>=0.5.1
python-doctr[torch]>=0.11.0
azure-ai-vision-imageanalysis>=1.0.0b1
deepl>=1.16.1
deep-translator>=1.11.4
stanza>=1.7.0
jaconv>=0.3.4
transformers>=4.49.0
diffusers>=0.28.0
wget>=3.2
largestinteriorrectangle>=0.2.0
paddleocr>=2.8.1
paddlepaddle>=2.6.1
loguru>=0.7.2
fugashi>=1.4.0
pyperclip>=1.9.0
unidic-lite>=1.0.8
fastapi
uvicorn[standard]
